+++ example/lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:declarative_navigator/declarative_navigator.dart';
 import 'package:example/src/examples/login_example.dart' as login_example;
+import 'package:example/src/examples/prevent_back_example.dart'
+    as prevent_back_example;
 import 'package:flutter/material.dart';
 
 void main() {
@@ -47,7 +49,12 @@ class _HomeState extends State<Home> {
 
   Widget get _exampleSelection {
     final examples = <String, DeclarativeNavigatorSource Function()>{
-      'Login/Logout + Child Navigator': () => login_example.MainAppNavigator(),
+      'Login/Logout + Child Navigator': () => login_example.MainAppNavigator(
+            close: _closeExample,
+          ),
+      'Prevent system back': () => prevent_back_example.MainAppNavigator(
+            close: _closeExample,
+          ),
     };
 
     return Scaffold(
@@ -69,5 +76,13 @@ class _HomeState extends State<Home> {
         ],
       ),
     );
+  }
+
+  void _closeExample() {
+    setState(() {
+      rootNavigator?.dispose();
+
+      rootNavigator = null;
+    });
   }
 }